Always send stats inline, never standalone.
Previously there were some cases where the connection code would need to send stats,
and it wasn't able to piggy-back them on a data packet.  A seperate standalone
stats message was used for this.  Changed to plumbing so that the connection code
could initiate sending of an SNP data packet, but ask SNP to not use the whole packet,
and pass through some data, back to the connection-type-specific code.  This way all
stats can be sent inline in a data packet.

This is a step towards getting rid of the LinkStatsTracker's acks.  Those are still
useful for stats on the relay leg of relayed connections, but for end-to-end, SNP
already does acks, and so it's silly and complicated to have two different mechanisms
for this.

Tweaked ack delay values and timeouts.  If the ping was realy low, then if a receiver waited
until the timeout and was forced to flush the acks (no piggy-back opportunity
presented itself), then the sender's timeout was exactly the same, and so there
was no room for error and the ack would just barely timeout.

Also when a packet is NACKed, spew a bit more detail about what actually caused the
nack (either an actual nack from the other side, or ack timeout).

Changed some constants related to PingLocation stuff.  That isn't really applicable
to this lib, but I've made some improvements to that system in the Steam code.