Big reliability layer fixes.
Robustify serialization of acks.  Previously we would just always send max of 6.  And if we
needed to truncate that (because of fragmentation due to high packet loss and latency), we wouldn't
include the timing information.  The loss of timing information meant that once this happened,
we wouldn't be able to provide RTT estimates, but those are used, among other things, for ack
timeouts.  So we could really go off a cliff.  Now if the ack flush timer has expired, then
we serialize everything that will fit in the packet, even if that means we cannot send data.
if the timer has not expired, then we first reserve a small, fixed amount of space for acks.  Then
we try to see how much data we want to send.  After planning out the data send, then we go
back and serialize the acks using whatever space is left, which might be more than we reserved
during the first step.  Also, any time we are able to send all of the pending acks, whether
because it has expired and we allocated as much space as necessary, or because we just happened
to have the space, then we clear the ack flush timer.  (If we only send the oldst acks, then we
don't clear the timer.)

Also fixed a bug with the receipt of reliable segments.  When we received a segment, we would
check if that completed an entire message, and dispatch it.  But if, e.g. we had [part of 5],
6, 7, 8 in the buffer, and we get a segment that completes 5, we need to dispatch not only 5,
but also 6, 7, and 8.  With the code we have before, we could never displatch more than one
reliable message per decoded reliable segment!

I found lots of problems by running the connection test with 200ms ping and 25% packet loss,
which are pretty terrible conditions.  But we need to make sure we don't go off a cliff.