Yet more refactoring to hook up plumbing in Steam.
Renamed CSteamNetworkingSocketsBase back to CSteamNetworkingSockets.
That's the only class we're going to have in the opensource code.
Also, don't do the thing where we derive CSteamNetworkingSockets
as the topmost concrete class.  As it turns out just doing some
simple upwards casts in a few places, kept the code a lot simpler
and made it more clear when shared code was accessing methods
that were only available in certain environments.

Switched initialization of the low level support routes to be an
explicit ref counted API.  This is complicated in Steam, and having
this will keep things easier.

Delete iclientxxx.h headers - found a better way to plumb this through.

Delete SteamNetworkingSocketsLib::g_eUniverse.  That is a thing that
only exists in Steam.