Update README.
Update roadmap section with note about  plans are for the current (FIRST!) release.

Delete comments about OpenSSL bloat.  That referred to a bunch of stuff that was being loaded in by opensslwrapper.  We still do have some cleanup to do here.  In particular, we'd like to be able to get OpenSSL to work for the 25519 stuff, and get rid of our reference code.  Another potential project is to use libsodium for everything, which may be leaner and faster -- however, that might require us to switch to AES-GSM.  That would be a fine idea, but it would introduce a wire incompatibility.  (We might end up doing this for the first release, to avoid having to deal with compatibilty issues.)

Also delete comments about STL container bloat.  I have converted almost all of the ordered maps to hashmaps, which do not do seperate heap allocations per item.  There are 2 or 3 remaining std::maps in the reliability layer, but those are constrained to a max size (to protect against a malicious peer), so I don't think this is high priority.  It would be good to fix this to limit the use of the heap to an absolute minimum, but for now tehre are more important issues.