Cleanup locking issues.
Delete unused mechanism to do all the background thread work from the main thread when callbacks were pumped.

Make SteamDatagramTransportLock::s_nLocked be a plain int, we don't ned atomic acces to this, since access
is gated behind the actual lock and only one thread may ever read or write to it at a time.  (Except if
an assert fails.)  Because this is accessed frequently, it's worth it to avoid anye xtra memory fence
operations.

Added a mechanism to loosen the threshold that triggers the assert for hlding the lock for too long,
when we run certain sections of code that we know might take a while.  This allows us to keep the threshold
tight for normal operations, which is really important for performance.

Delete unnecessary taking of the lock in OpenRawUDPSocketInternal, we assume that we already hold the lock.

Tweak assert that checks for leaked sockets when shutting down lowlevel systems.