Use AES-GCM for per-packet encryption.
Also tweak the way that the per-packet IV is constructed.  Previously,
64 bits were the packet number.  This meant that the attacker knew 64-bits
of the IV, while the other 64-bits were secret, generated by key exchange.
Now, the IV is adjusted by the packet number, so that all 128 bits are
unknown to an attacker, although he does know that the relationship
between one IV and the next.

This introduces a breaking wire incompatibility, so I bumped the protocol
version number as well as the minimum required version for the peer.  We
still have luxury of doing this sort of thing, but we won't once we make
the first release.  We'll need an exceedingly good reason to break backward
compatibility with older peers.  (Although, for example, if a security
vulnerability is detected, we may allow the application to control whether
it wishes to allow conneections from peers using the version with the
vulnerability.)

Fixes #43