avoid memory leak caused by vtable mismatch
In the main GNS project we had VALVE_CRYPTO_25519_OPENSSLEVP defined,
but in test_crypto we did not. This pointed out a serious design issue,
in that the vtables were not matching up and te wrong destructor could
be called if a macro was/wasn't defined.

This is ugly, and we should probably use the C++ pimpl idiom here, but
this works well enough and we won't have both behavior types possible in
the same build of the library (it's either donna or it's OpenSSL). So
there's no reason to add the extra complexity of an abstraction layer
for the class internal structure.

Signed-off-by: Steven Noonan <steven@valvesoftware.com>