CEC25519KeyBase derived from CCryptoKeyBase_RawBuffer again.
(Partial undo of 63aec987cd1b7d7c54fdde5e80f0f0dd64962579.

I'm going to need CCryptoKeyBase_RawBuffer in Steam.  But now we *always* derive from CCryptoKeyBase_RawBuffer instead of doing so conditionally based on a #define.  Those #defines are not available to the application; it doesn't know how the library was compiled.  I think the standard way to solve this is with a configure script, but in this case I think it's easier to just make the ABI for the class consistent, and waste a tiny bit of CPU time and memory.