Use 96-bit AES-GCM IV's.  Secure the protocol number.
Switched AES-GCM IV's to be 96-bit instead of 128-bit.
- This is what TLS uses (RFC5288)
- This is what NIST recommends (https://dl.acm.org/citation.cfm?id=2206251)
- This is more optimal.  (Certainly for OpenSSL's implementation; I believe it's true in general for GCM.)
- This is the only value that libsodium supports.  Note that we're still using a 32-bit tag, and libsodium only supports the full 128-bit tag (to my knowledge, see #48).  But we might decide to use the full 128-bit tag, and that that point libsodium could be used.

Moved the protocol number out of transport-specific handshake packets to be a "crypt parameter", where it will be signed.  In the future we might need to allow some clients to connect to older clients that have a known security vulnerability (perhaps only during some temporary transition period).  We don't want an attacker to be able to forge the protocol version and force the weaker protocol to be negotiated.

Bumped the protocol version because these are wire-compatibility-breaking changes.

Added a specific error code if a connection fails due to protocol incompatibilities.