Be much smarter about when to send acks and nacks.
Previously, we had a single global timer that controlled when we needed to flush out all acks (and nacks).  Now, we actually track a seperate deadline to ack each block of received packets and when to nack each block of dropped packets.  This data was added to the existing packet gap map.  But the logic got significantly more involved.

One big problem with not having two different timers is that ACKs really cannot be delayed beyond their due date, or else the sender will retransmit unnecessarily.  They have a much longer delay (currently 100ms), and so we are almost always have the bandwidth available to send one, if we haven't piggy-backed on another outbound packet before then.  NACKs, on the other hand, although we do wish to send them as soon as possible, have a much shorter delay (now 3ms, was 10ms).  But if necessary, we *can* wait until we have the bandwidth to send an ordinary packet.  What was happening previously is this: we would be sending as fast as the bandwidth limit would allow, and then we would detect a dropped packet inbound.  This would require a NACK in 10ms, and we would set the generic timer.  When that timer exipred, we would be over the bandwidth budget.  We didn't know it was for a NACK, and so we could wait, we just assume that the timer expired, so we need to report on all packets.  This cuased us to hit the "this should really never happen under normal conditions" code, where we would spew warning text, and then send a packet with nothing but acks/nacks in it.

Now the logic for deciding when we want to send a packet is smarter.  If we have acks due, we must send them, even if we are over budget.  Nacks, however, are treated the same as data send (first time or retry), thus they are subject to the bandwidth limit.

Furthermore, when reserving space for acks while laying out a packet during sending, we figure out which acks *really* need to go out.  We not only send the acks that must be sent now, but we also look forward into the future, up until the next time we will have bandwidth to send a packet after the current one, and we send all acks that we would want to send before then.  Finally, if we ever do end up being over budget and needing to send a packet with nothing but acks, we go ahead and flush all the acks/nacks that are ready, not just the ones that really muct go out now.  It's almost always just a few more bytes, and getting all caught up could make a big difference.

All of this sophisticated bookkeeping doesn't really kick in or add any extra work when packet loss is low, since the number of gaps is very small.  When packet loss is heavy, being smart pays off.  We end up sending about 10% fewer packets, and the effective reliable RTT experienced by the app (e.g. app sends a reliable ping, and then sends a reliable reply) also drops.

Also, the spew about using all of our bandwidth just sending stats and acks does not fire anymore.

I also added a method for "logging" every outbound packet into an internal vector.  This is way more useful than actual test logging, because it's really fast, and also you can collect everything at different points in the code, but they appear together.  Doing this with printing is really difficult, and the output is a lot harder to read, and it's a lot more likely to impact the speed and thus alter the behaviour of what you are tyring to test.  (E.g. standard problem where your profiler is slow.)

Also fixed a bug, when reserving space for acks, we were not adding the extra byte needed for the explicit count byte when there are more than 6 blocks.  This would mean the last one would never fit.