Refactor how we populate the inline "stats".
To send a data packet, first we ask the transport-specific code to
prepare a context, and figure out how much space it needs to reserve.
Then it passes that into the generic SNP code, which lays out the packet.
Then SNP calls the transport code to actually send the encrypted data,
providing the context that the transport layer originally provided.

This simplified a bunch of stuff and allowed a bunch of cleanup
(especially in the code for relayed connections -- not visible here).
Most importantly (and the reason I realized I needed to refactor this),
it facalitated fixing a serious P2P transport bug, that would have been
awkward to fix without the refactor.