Delete a bunch of CPU info stuff.
We only needed to know if SSE2 support is present, only in 32-bit, and only
if you are using our 25519 code, which most people hopefully will *not* use.
(Use OpenSSL.)

To answer this one question in this one place, we were using a whole framework
to get a ton of other info, and all of the code had ben ripped out of Steam
and was really gross.