Make LinkStatsTracker::SetDisconnected virtual.
In a few places I need to call this through a base class pointer in
the Steam relay code.  Ths stats tracker heirarchy is really complicated.
Might be nice to clean it up and find a way that we never do virtual
function dispatch when we could have just inlined code.