Implement a rudimentary PKI.
Store a list of public keys and certs we have for them, and implement logic
to evaluate their trust.

New cert checking flow in the connection code.  I found a *tiny* bug where
we were not actually checking the signature of the connection crypt info
using the signature in the user cert (!).  We were checking to make sure that
the user cert was valid, but anybody in the middle could have substituted
their own ephemeral keypair for key exchange, and we wouldn't have known.

Certs can now be issued to multiple AppIDs.

Cert tool can now emit json output and bind certs to particular AppIDs.