Fix bugs not properly piggy-backing tracer pings.
We were setting the flag too agressively, and not really allowing for any
opportunism.

Really, tracer pings in stats messages are not that useful end-to-end, and I
probably should delete them, because every single data packet has acks that
serve as ping measurements.  If the conversation is very thin, it is perhaps
possible that you could go a while without being able to use the acks for
latency measurement, and tracer ping requests would kick in.  But it's just
as likely that a plain old keepalive would end up be sent, too.

Tracer ping requests are much more useful for maintaining a latency
measurement with the relay, on relayed connections, because there is nothing
that happens constantly due to normal traffic that can be used to measure
latency with the relay.