Some improvemenets to CUtlMap.
Inline a bunch of stuff in the internal implementtion.  This can make a big
difference in the debug build.

A much simpler, saner iterator implementation.  We don't need full STL
iterator functionality.  We just want range-based for loops to work.  And we
don't want to expose a non-const reference to the key!!!!!

And some other stuff from Steam branch.