Fix bug waiting longer than caller requested.
This was a problem for manual polling mode.

Also for "normal" mode, where there is a service thread, increase the timeout.
We should never depend on this timeout to avoid any bad performance or solve
race conditions.  We should explicitly wake up the trhead when needed.  So if
we have a bug, we want it to stall for a long time, and cause people to ask
questions and investigate.