use unreliable when no fragmentation packet < MTU and reliable otherwise
if we can send in one packet with stats+encryption+ipv6 header + payload that means payload should be 1132 or less, we can send in one packet unreliably because we will have multiple peers sending as well, chances are it will arrive. For fragmentated packets we take advantage of Valve's reliability protocol.