Refactor base class connection state machine.
Make explicit hooks for the connecting and finding route states.

Having the base class always in charge of sending connect request messages
and retrying, and making assumptions about what role the transport has in
doing those initial messages just didn't map well to what we want the P2P
code to do.  It needs to be doing a bunch of different things at once.  There
was probably a way to implement this all within the catchall hook
ThinkConnection(), but that was too complicated to understand.  The nature
of this code, the dereived class already has to have a pretty detailed
understanding of the base class implementation, which is not ideal.  That
direction would have made it even worse.  With this direction, the hooks
are a bit more clearly defined.

Also refactor P2P inititlaization a bit, moving some duplicated code into
CSteamNetworkConnectionP2P::BInitP2PConnectionCommon