More progress on WebRTC connections.
Need to pass the transport to SNP_SendPacket.  We cannot assume that the
current transport is the one who filled out the context and is sending
the packet.

Got some signaling messages working to send answer/offer and candidates.
This is probably more complicated than neccessary, but it is because I
want to assumevery little from the signaling framework.  Specifically,
it is *not* assumed to be reliable, and I assume that really large messages
might not work.  I do not currently assume that messages will gitin a UDP
packet, although limiting signals to <2K or so seems like a good idea.

The WebRTC connections are managing to get connected.  Now I need to have
logic to update the SteamNetworkingSockets state machine, and if we have
SDR, I need to intelligently decide when to use the NAT-punched route
versus the relayed route.  I'm expectig that the relayed route could be
faster in some cases, although I don't know how often.

I am pretty sure that I am using too much WebRTC, and all I really want
is the ICE NAT punching bit, and this will eventually get rewritten.