Alway use VALVE_CALLBACK_PACK_LARGE
Steamworks SDK has found itself in a really bad place having to support ABIs
that differ on different platforms.  That didn't matter in the days when
everybody compiled all of their binaries per platform, but it breaks the
"write once, run anywhere" ideal.  We don't have to support old ABIs so let's
just pick a packing and use it everywhere.  (But note that we do have stucts
with pointers in them, so our structs will vary between 32-bit and 64-bit.)

https://github.com/nxrighthere/ValveSockets-CSharp/issues/8