Move rendering helpers.
When linking with Steam, these call ISteamNetworkingUtils functions, and
GCC complains if it hasn't seen that class.  Even if you don't ever call
any of these functions, GCC doesn't like the fact it could not generate the
code if you were to do so.  Not sure what the spec says, but it feels like
Visual Studio's approach makes more sense here.