Add ETW instrumentation
We've been using ETW to track down performance issues with P2P connections.
The Steam link use case target is 50Mbit, and we need consistent low latency.

It's not enabled in this code, but I don't want to have to deal with merge
conflicts.  With just a tiny bit of work, it could be made to work in this
lib.  I'll probably end up yanking some of the webrtc events later, because
those required hacking webrtc code, and i want to be able to use webrtc code
without any modifications at all.  We think we have found a way to use webrtc
code in a performant way, at least on machines with more than two cores.  On
lower end hardware (e.g. the Steam link hardware), we really need to avoid
context switching and minimize the number of threads, and the webrtc
architecture will probably not work.  Might need to see what google did for
Android.  It might be more appropriate.