Plumb through proper STUN credentials.
The "user fragment" is the connection ID, which is already being shared.
And the password is a random number.  Both of these values are Base64 encoded.

Changed the server to be the controlling agent.  This is opposite of the ICE
default convention, but ICE worksjust fine either way.  The reason I want to
have the clientbe the controlling agent is that this same role assignmehnt
will apply to the higher level routing decision of whether to use the ICE
route or relay over SDR.  And in themost immediate use case I am working on
(steam remote play), the server will have some convars available that might
notbe accessible on the client (e.g. a phone).  This is kind of arbitrary
and maybe not optimal for future use cases.  I could plumb thorugh some
options to control this later, but for now this is an OK default.

Changed the way that initial ICE settings were plumbed through.  Pass them
in a struct to the factory func, instead of making them callbacks on the
delegate.

Changed the way that candidates were sent back and forth in signals.  Used a
more generic reliable stream of messages instead.  Also made the receiving a
bit more robust, if a message contains some retransmission and also new
data.  (Previously it would ignore everything if the first message was not
the next thing expected.)

Set webrtc network thread priority.