Refactor rendezvous messages.
Promoted concept of a stream of reliable rendezvous messages to the
connection layer.  That is not a thing that only ICE might find useful.
(Although right now, only ICE is using it.)

Manually base-64 encode the ICE credential values, and only use the
lower 30 bits, so that we get exactly 5 characters, without any '='
padding.

Added IsControllingAgent(), which has been promoted to a connection layer
The terminology is from ICE, but the concept is generally useful.  I will
use that concept when deciding which transport to use (ICE or SDR), to
avoid asymmetric routing.

Fix bug where we might try to send a queued reliable rendezvous message
before we know who we are sending it to.

Added code to take appropriate action when ICE tells us the writable state
has changed.  Previously we were relying on polling, but now we definitely
need something to kick-start the thinking loop.

Added some hack code to try to load up the steamwebrtc library manually,
if we have no CreateICESessionFunc function already.

Refactored P2P rendezvous handling a bit, moving more of it into a member
function.

This is a wire-protocol-breaking change.  I didn't bother doing any
versioning because that is not necessary yet.