Fix another subtle bug when filling a gap.
If we were scheduled to ack the packets before the gap, then we must not
forget to do this.  When we delete the gap, we must propogate that deadline
to the next gap (or the sentinel).  If we don't do this, then we will either
totally forget to schedule an ack (if nothing after the gap requires acks)
or we will do it late (since the ack time for later packets will usually be
later).

This is not a protocol breaking bug, because SNP always acks everything any
time anything is sent.  So messages still get delivered reliably.  And if
there is any traffic (e.g. a game sending packets et 30hz), odds are that
these acks are getting flushed out anyway.  But it does mean we could be
relying on timeout on the sender side, when the receiver really needs to
be sending the acks in a more timely manner.