Refactor to support ISteamNetworkingMessages.
Previously ISteamNetworkingSockets was plumbed through much more invasively.
But with the changes to callbacks, symmetric ocnnect mode, and P2P
connections automatically using internal loopback when sending to self,
now it is possible to implement ISteamNetworkingMessages much more as a
seperate layer on top.

ISteamnetworkingMessages is a P2P interface that is more like UDP.  You don't
listen or connect, and you don't deal with connection handles.  Instead, you
specify the remote address with each send call.  The purpose of the API is to
make it easier to port UDP code to P2P.  It's not currently part of the
opensource code, and has not yet been released in the Steamworks SDK.  But I
probably will soon, once we have some good P2P example code.  The previous
incarnations of the interface were kind of crap.  I finally have something
worth releasing.