Handle signals with no 'to' connection ID better.
Move processing of connection_closed signal to
CSteamNetworkConnectionP2P::ProcessSignal.  This is because in some cases,
the peer might start to connect, and then cancel.  In this case, the
connection closed message will not have a "to" connection ID, since they
don't know ours (yet).  So we must locate the connection, if any, using the
from connection ID.

Also, added some comments about why it is important to not reply if we do
not recognize the connection.