Process reliable segments in the finding route state.
If we ack these packets, the sender will assume that we have received them!
This fixes a bug, currently exhibited by test_p2p, where we queue reliable
data immediately after attempting connection.  It does properly get sent in
the very first end-to-end packets.  Then the receiver drops the segments,
while acking the packets.  This leads the sender to assume that the data
segments have been processed, and it will not retry.