Simplify and optimize CUtlHashMap
These changes reduce the number of times that we have to probe back into old
buckets when the table has grown and not all the items have been hashed into
the proper bucket based on the current table size.

(For context: this hashtable does not immediately rehash the items when the
table needs to increase in size, since that leads to a spike in CPU time.
Instead, the cost of cleanup is amortized.  For large tables, this can be a
really big deal.  Maybe for the relatively small tables that are likely to
be encountered in this library, this amortization is not that important, but
then again this code is intended for realtime use so reducing the worst-case
perf is important.)

Probing back to check more than one bucket is relatively costly, since it
means touching more memory pages.  Ideally, we can do the hash and jump
straight to the item we are looking for.

A few basic observations are behind the changes:

- For many common use cases, doing the cleanup only on insertion is not
  enough.  For example, consider a use case where many insertions are done to
  initialize a lookup table, which is then never subsequently modified.  Such a
  table will go its entire lifetime in an impaired state.  But even if the
  table does have inserts, there really need to be a sufficient number of
  deletions so that the cleanup sweep can catch up.  If there are no deletions,
  in fact, the cleanup sweep will end *right* around the time that its time to
  double the table again, and we have to start all over!  (See below for more.)

  In a use case in Steam unit test runs, doing an incremental rehash on Find()
  reduced the number of probes into old buckets by a factor of ONE THOUSAND (!!!).
  This depends significantly on the size of the table and the access pattern,
  obviouisly, but the main takeaway I think is that relying only on cleanup to
  hapen during insert can lead to some pathological cases.

- Under normal cirucmstances, where the table size doubles only as a result
  of insertion, you should only have to probe back *one* bucket.  (Even if you
  are only doing cleanup on Insert.)  Consider this sequence of events:

  1. Table is clean with N items.  (Note that table size is always a power
     of two.)
  2. An insertion causes the table size to double to 2N.  Buckets 0...N-1 need
     to be rehashed.  Note that at this point, only one back probe is necessary
     We know that the item will either be in the bucket it would be in with its
     current size of size 2N, or the bucket appropriate for a table of size N.
     But no further back probes are needed.
  3. N more insertions happen.
  4. At this point, even if the absolute worst case (every bucket has an item
     in it, which means there were no hash collisions, and also no additional
     cleanup has happened because of Find() or Remove() ), the sweep completes
     and the table is clean *just* as the table gets full with 2N items.  In
     practice, the sweep will almost always happen before the table needs to
     double.
  5. Now we are back where we started, at step 1.

  However, if app code calls EnsureCapacity(), the table may double more than
  this.  So, we keep track of the smallest table size that was ever used to hash
  any extant items into their current bucket.  This handles the case where the app
  increases while a sweep is in progress, in which case we would actually need
  to probe back more than one time.  It also means that if our analysis above
  was off-by-one, it doesn't matter.

  The previous code would probe back up to log2(N) times!  If N=1mil, which is
  totally a size we encounter in backend server code all the time, that's 20
  probes!  Of course the average case will be much less than 20, since after
  the first 1000 items have been cleaned up, which is only .1% of the table,
  the number of probes has been reduced to 10.  But still, it's a lot of back
  probing that we can easily prove is not necessary.

With those changes, keeping track of the bitfield of which buckets have been
hashed is much less valuable.  My tests do show that it still does reduce back
probing somewhat, but I don't think it's worth the cost of maintaining and
checking bitfield in the first place.  And in this library, where table sizes
are probably likely to be relatively small, its of even less value.  Also,
removing the bitfield allows me to reduce the number of memory allocations
and code dependencies.  Deleting code is always a win.

This change also includes a number of micro-optimizatiosn intended to make the
code smaller, especially debug code.  It might not be a big deal on a relatively
small library like this, but on a big codebase like Steam, it's hundreds of Kb
reduction in the size of the .dll, less work for the compiler and linker and
optimizer.

P4:6176888