Move platform specific stuff to seperate header.
We are starting to include steamnetworkingsockets_internal.h in quite a few places.
This is bringing in Windows.h and a ton of stuff that most files don't need.
Segregate that mess to a header to only be included in the few files that actually
deal with low level sockets.