Fix P2P scheduling bug related to processing timeouts and anti-flap checking.
CSteamNetworkConnectionP2P was scheduleing a wakeup call EXACTLY at the same
time as the timeout.  But it was not actually responsible for processing the
timeout, the transport does that.  The transport did properly have a method
scheduled to deal with it -- but they were both using the same wake up time.
So the connection kept waking up first, and scheduling the same wake up time
as the transport, and then winning the tie, repeatedly.  The fix is to make
sure that we wake up *just* after the transport has dealt with the timeout.

It's possible that the thinker system could deal with these sorts of cycles
better, but I'd rather avoid adding that complication until I find a problem
that doesn't have a simple solution like this.

P4:6653586