Change "deadline < usecNow" to "deadline <= useNow" in several places.
We often schedule a wakeup *exactly* at the deadline.  The OS almost never
wakes us up at *exactly* the right time.  But when it does, we can get
confused, setting a deadline for the current time, and then not taking
action on the thing we requested the wakeup call for, and setting the
deadline again.

P4:6653641