More progress on multiple lanes
Changed my mind on the lane select command.  Encoding this as a delta
from the previously selected lane really only gives a tiny benefit,
a single byte *if* several stars align.  And even then, that savings
could not be predicted in advance.  Furthermore, doing this required
sorting the lanes, which is just bad.  So basically it added a bunch
of runtime complication for extremely dubious benefit to the wire
encoding.  Just a bad tradeoff all around.

So, new design: the lane select command always *sets* the current lane.
Much, much simpler for encoding, and almost no loss to wire efficiency
except for a byte or two in some edge cases.

I did keep one small optimization, which is that lane 0, if present,
will always be encoded first.

Still a work in progress.

P4:6818528