Better handling of relocatable types.
Separate out "trivially copyable" and "relocatable" (which I guess is
the same as "trivially movable".

By default relocatable will be the same as trivially copyable,
however since this is not in the std::namespace, I am explicitly
creating an opportunity to tag your type as being relocatable.

Because this makes it possible to do stuff GCC doesn't think is
safe, we'll need to turn off their warnings.