Fix issues with refcounting of reliable segments.
If our ping is very high and send rate is low, we can easily receive ACKs
for all outstanding reliable data of a message we have sent, while it is
still the first item in the queue and we still have more to send for that
message.  If this happens, we would decrement the last reference count and
delete the msssage, even though we haven't finished sending it.  To fix
this, just have the queue hold a reference to the message.

P4:6868998