Don't use htonl, etc.  Use BigDWord and friends.
Yes those are weird and non-standard.  But on Windows ntohl actually is a (dll
import) function call.  This is slower than our defines and inlines (probably
not a huge deal).  Also, in situations where we just want to parse/print IP
addresses we need to link with winsock?  That is dumb.