Implement socket polling using epoll() where it is supported.
This is a fine change and all, probably a small optimization.
The real reason to do it is to make the PS5 port easier.

I have to admit that it totally worked the very first time, which
is kind of scary -- obviously something is deeply broken.  If so,
it is possible to switch back to regular polling by just defining
USE_POLL instead of USE_EPOLL

P4:7194744,7194845,7194874