Improved platform compatibility layer
Deleted several places where we were using this pattern:

   <platform 1 code>
   <platform 2 code>
   ... etc

This is a pain when you have some platforms that are under NDA and you
cannot share the code.  Instead I introduced some abstractions that made
it possible to move all of the platform-specific code (at least for NDA
platforms, which are my primary concern right now) to a separate file.
This way, we can just exclude particular files from distribution, rather
that have files that actually differ.

Also changed some platform-defines to more "functional" defines (USE_POLL,
USE_EPOLL, WAKE_THREAD_USING_EVENT, WAKE_THREAD_USING_SOCKET_PAIR) that
make it clear the reason for the #ifdef.  This also means on platforms
that support multiple methods (e.g. Linux) we could switch those options.

Finally, renamed a few things with "SteamDatagram" in the name to
"SteamNetworkingSockets".  I'm trying to have "datagram" refer exclusively
to the relay network, which is only available to Steam partners, and
thus in general should not be present in the opensource code.

P4:7219053,7219056,7219071