Fix ICE/P2P cmake stuff
ENABLE_ICE means "enable any ICE support, including the native client"
and is on by default.  (Because the native client is not much code.)

Try to build the trivial_signling_server if we are building either
examples or tests, and ENABLE_ICE is set.  The P2P test needs it.