Fix for rare deadlock with P2P callbacks.
When looking up connections by their handle, take both the table and the connection
lock using a timeout.  If we fail, release both locks and start all over.
This will proptect against a deadlock if we take the locks in the opposite order.
We currently don't ever encounter this situation solely within this library.  However,
we did come across a case with P2P callbacks calling into a custom signal handler.
We could probably fix the custom signal handler to not do this, but it seems like
there are going to be other edge cases, and this change is low risk because contention
is low and most lock attempts will succeed immediately.  And deadlocks are very
hard to reproduce and debug.

CR:saml

P4:7307214