Automatically fix certain trivial out-of-order messages conditions.
When queuing a message, check if higher numbered messages are already in the
queue.  If so, then we can silently fix this up and hide this anomaly from the
app!  This will be a common sitution when packets arive in quick succession
out of order.  I am planning to also handle this at the packet level (which is
useful so I can tell when the situation is causing real problems at the link
layer and when it is being corrected).  However there might be certain
misorderings that remain, even after this which we can stil fix, for example
if they arrive out of order but after a delay, or if we have to retry.
If the host is only polling for messages every 30hz, we may have plenty of time
to correct the situation.

Many games do not handle out of order packets at all, and simply drop them,
so fixing this up automatically directly improves the connection quality.

P4:8088078