Change how variables related to config options get named.
They were previously named with some prefixes that made it awkward to find all
of the uses with a naive text search with "match whole word only" on (which is
how Entrian search works and most other searches that do indexing).

Now, instead of name prefixes, I put them in a namespace.  I also ditched the
"m_" hungarian thing on the ConnectionConfig members

P4:8312403